Address bug #36, effective LAI, total SAI and effective SAI were being zero'd strangely in ed_clm_link.
It appears that this part of the code is responsible for zero'ing the bare-ground portion of the code.  There is also the
assumption that patches that are not within the perview of ED should be zero'd here too. This would be crop patches, as well as the
bear-ground patches (grrr) just mentioned.