Workaround for ed / clm-r153 snow bugfix restart problem
Merging ED with clm4_5_6_r153 caused ED bit for bit restarts to fail
for f10 and f45 resolutions when they had previously been passing. The
delta in r153 that causes the failures is the change in inequality in
CanopyHydrology, snowment >= 0 to snowmelt > 0. This had previously be
the default in clm, but an eariler clm tag changed it from '>' to
'>='. r153 changed it back. But there had been other changes to snow
in the intermediate tags.

This commit simply reverts the inequality change so merging to ED to a
modern clm trunk tag can continue. It should NOT be merged back into
the clm trunk!

More details are in NGEE-T ed issues: #46, #74

User interface changes?:

Code review: self

Test suite: ed - yellowstone intel, gnu, pgi
Test baseline: not run
Test namelist changes: n/a

Test answer changes: change in science logic, so potentially not bit
for bit, scale of changes unknown.

Test summary: pass except for expected failures in f09, f19 (#14) and
gnu f10 (#43),