Bugfix for nag: dangling pointer to fates sites
Creating a pointer to an object, such as currentCohort%siteptr, requires
that the target be defined as a target or pointer. Since sites are a
dynamic array inside a derived type and 'allocatable, target' doesn't
appear to be a valid variable attribute, the only way I see to make this
work is with change fates sites to a pointer.

Tests: SMS_D_Mmpi-serial_Ld5.5x5_amazon.ICLM45ED.hobart_nag.clm-edTest
gets past the danging pointer runtime error.