Effect of active crown fire on the Rate of Spread
Previous commits computed the occurrence of active crown fire and set
currentCohort%fraction_crown_burned = 1.0_r8

Here I begin to introduce the effects of active crown fire on the size
and intensity of the fire as discussed with @jkshuman in #573. First
step in this is the effect of active crown fire on the Rate of Spread.

I have confirmed that, with the code changes this far,
history varialbes FIRE_ROS, SUM_FUEL, fire_fuel_bulkd have changed,
while fire size and intensity have not changed. This is what I expected.