[In EDCohortDynamicsMod increase maxCohortsPerPatch to 300 if cohort age tracking is on ]
[If cohort age tracking is on then cohorts can only fuse if size and age
are similar enough. This means we end up with many more cohorts and maxCohortsPerPatch
needs to be increased. I added a local variable maxCohortsPerPatch_age_tracking to
EDCohortDynamicsMod.F90 that is used in place of maxCohortsPerPatch if
age tracking is on. Not sure this is the best way to go about it but it seems to be
working.  ]

Fixes: [NGT-ED Github issue #]

User interface changes?: [Yes (describe what changes), No]

Code review: [Names]

Test suite: [suite name, machine, compilers]
Test baseline:
Test namelist changes:
Test answer changes: [bit for bit, roundoff, climate changing]
Test summary: Quick test on lobata. It compiles and runs.