HLM harvest data passed to FATES
The harvest data from landuse.timeseries are passed to FATES for
logging. The FATES logging parameters are used to guide harvest.
Only area fraction data are implemented, and these fractions are
applied to the FATES logging intensity parameters. The desired
effect is for the given logging intensity to be applied to a fraction
of the cohort determined by the HLM harvest rate. Separate fractions
are applied to primary and secondary patches. This is a working version
for testing.
 Changes to be committed:
	modified:   biogeochem/EDLoggingMortalityMod.F90