Updated fates logging calc to account for all secondary land creation
Conversion to secondary land due to harvest is based on the area
harvested, not just the logged crown. Also, the harvest_carbon_flux
diagnostic has been updated to align with the wood_product output.
Note that these are in different units and that monthly average
output during harvest month is different for each.
 Changes to be committed:
	modified:   biogeochem/EDLoggingMortalityMod.F90
	modified:   biogeochem/EDPatchDynamicsMod.F90