Grass PFT only creates leaves not cwd
In response to issue #884

Modified EDCohortDynamicsMod.F90 (@ln970): if the pft is grass then all of its organs are treated as leaves for calculation of coarse woody debris(cwd).
EDPatchDynamicsMod.F90 (@ln 1641):  When patches spawn, only tree pft donate dead tress or cwd, and mass to the new patch, if pft is grass it assumed that grass was consumed in fire. This may not resolve fluxes to the atmosphere accurately.
EDPhysiologyMod.F90 (@ln 2199):  When other litter fluxes occur, if pft is grass, then it only turnsover  or contributes leaves not coarse woody debris.