Update xml patch file for api25.1.1
This commit creates two xml patch files to facilitate
this update.  The file names are to be applied in the
order denoted by the number suffix on the xml file name.

The first file is used to create the new Atkin parameter
as well as update the mortality rate typo.  The UpdateAPI.py
tool is modified to allow for variables to be added that
are undefined.

The second file updates the new Atkin parameter using
BatchPatchParams.py.  The update intentionally leaves
the final pft as undefined.

Modifying and creating new variables with mixed values,
in this case leaving one undefined, is what necessitated
the update to the tooling and the sequential tool calls.