Fix BatchPatchParams.py global history update call
The call to update the global history attribute of the
output file was missing a further call to the sort tool.
This was fixed by moving the history update to earlier
in the code.

Note there is still an issue in which the resulting
parameter file cdl results in an incorrect name due
to the use of chained temp files in the tool calls.