update patch file names to include only pull request numbers
Removing the date strings from the patch files as I think it doesn't
really add anything to help the user.  I also update the base file in
multiple xml files to use the fates_params_default.cdl so as to build on
the previous xml files included in this pull request.