On patch creation, copy fire info from youngest patch (if any).
This minimizes pollution of the outputs from new patches. It's better than using the default values for new patches, but it would be better for new patches to actually get their fire weather stuff updated between creation and history output. Note, however, that since Nesterov Index is an accumulated sum, it will need to take an initial value from something. Otherwise we'd basically be saying that fuel is just as wet after burning as it is after rainfall.